+++ lib/jsoner.rb
@@ -5,10 +5,26 @@ require "jsoner/table_factory"
 require 'nokogiri'
 
 module Jsoner
+
+  class NotFullTable < StandardError
+  end
+
   class << self
 
     def parse(html)
-      Jsoner::Table.new(Jsoner::TableFactory.new(Nokogiri::HTML.parse(html))).to_json
+      html = filter(html)
+      if factory = Jsoner::TableFactory.check(Nokogiri::HTML.parse(html))
+        Jsoner::Table.new(factory).to_json
+      end
     end
+   
+    def filter html
+      if File.file? html
+        File.open(html) { |file| file.read } 
+      else
+	html
+      end
+    end
+    
   end
 end