+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## v0.0.2
 * Filter other elements beyond table element
 * Get first table element when having more than two in parsed HTML
+
+## v0.0.3
+* parse HTML file including HTML Table
+* fixing file is empty or have no table element